requirements.txt
flask==3.0.2
flask-pymongo==2.3.0
python-dotenv==1.0.1
flask-cors==4.0.0
bcrypt==4.1.2
PyJWT==2.8.0
python-jose==3.3.0
email-validator==2.1.0.post1
python-dateutil==2.8.2
geopy==2.4.1